에너지get float으로 바꿈. 대시잔상 메시에 맞춤

+++ Assets/01_Scripts/20_InGame/Managers/ObjectsManager.cs
@@ -20,7 +20,7 @@ public class ObjectsManager : MonoBehaviour {
   public float minSpawnInterval = 8;
   public float maxSpawnInterval = 12;
   public int cubesByEncounter;
-  public int energyGets;
+  public float energyGets;
   public int numSweetsPop;
   public bool isNegative = false;
   public bool destroyWhenCollideSelf;
